header

+++ template/header.php
@@ -50,26 +50,38 @@ if (isset($_SESSION['role'])) {
     <nav class="navbar navbar-expand-sm justify-content-center bg-light px-5 py-3">
         <div class="container-fluid p-0 justify-content-sm-end justify-content-center">
             <ul class="navbar-nav">
-                <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/">รายการปริญญานิพนธ์</a></li>
-                <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/thesislistwaiting">รายการที่รอตรวจสอบข้อมูล</a></li>
-                <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/thesisadd">เพิ่มข้อมูลปริญญานิพนธ์</a></li>
-                <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/thesisdelete">ลบปริญญานิพนธ์</a></li>
-                <li class="nav-item text-center dropdown">
-                    <a class="nav-link dropdown-toggle pe-0" href="#" data-bs-toggle="dropdown">จัดการสมาชิก</a>
-                    <ul class="dropdown-menu dropdown-menu-end">
-                        <li><a class="dropdown-item" href="/FinalProj/officeradd">เพิ่มข้อมูลเจ้าหน้าที่</a></li>
-                        <li><a class="dropdown-item" href="/FinalProj/temporaryadd">เพิ่มข้อมูลเจ้าหน้าที่ชั่วคราว</a></li>
-                        <li><a class="dropdown-item" href="/FinalProj/teacheradd">เพิ่มข้อมูลอาจารย์</a></li>
-                        <li><a class="dropdown-item" href="/FinalProj/studentadd">เพิ่มข้อมูลนักศึกษา</a></li>
-                    </ul>
-                </li>
-                <li class="nav-item text-center dropdown">
-                    <a class="nav-link dropdown-toggle pe-0" href="#" data-bs-toggle="dropdown">สถิติข้อมูล</a>
-                    <ul class="dropdown-menu dropdown-menu-end">
-                        <li><a class="dropdown-item" href="#">สถิติการจัดเก็บเล่ม</a></li>
-                        <li><a class="dropdown-item" href="/FinalProj/chart_director">สถิติการกำกับเล่ม</a></li>
-                    </ul>
-                </li>
+                <?php if (isset($_SESSION['role'])) : ?>
+                    <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/">รายการปริญญานิพนธ์</a></li>
+                    <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/thesislistwaiting">รายการที่รอตรวจสอบข้อมูล</a></li>
+                    <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/thesisadd">เพิ่มข้อมูลปริญญานิพนธ์</a></li>
+                    <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/thesisdelete">ลบปริญญานิพนธ์</a></li>
+                    <li class="nav-item text-center dropdown">
+                        <a class="nav-link dropdown-toggle pe-0" href="#" data-bs-toggle="dropdown">จัดการสมาชิก</a>
+                        <ul class="dropdown-menu dropdown-menu-end">
+                            <li><a class="dropdown-item" href="/FinalProj/officeradd">เพิ่มข้อมูลเจ้าหน้าที่</a></li>
+                            <li><a class="dropdown-item" href="/FinalProj/temporaryadd">เพิ่มข้อมูลเจ้าหน้าที่ชั่วคราว</a></li>
+                            <li><a class="dropdown-item" href="/FinalProj/teacheradd">เพิ่มข้อมูลอาจารย์</a></li>
+                            <li><a class="dropdown-item" href="/FinalProj/studentadd">เพิ่มข้อมูลนักศึกษา</a></li>
+                        </ul>
+                    </li>
+                    <li class="nav-item text-center dropdown">
+                        <a class="nav-link dropdown-toggle pe-0" href="#" data-bs-toggle="dropdown">สถิติข้อมูล</a>
+                        <ul class="dropdown-menu dropdown-menu-end">
+                            <li><a class="dropdown-item" href="#">สถิติการจัดเก็บเล่ม</a></li>
+                            <li><a class="dropdown-item" href="/FinalProj/chart_director">สถิติการกำกับเล่ม</a></li>
+                        </ul>
+                    </li>
+                <?php else : ?>
+                    <li class="nav-item text-center"><a class="nav-link" href="/FinalProj/">รายการปริญญานิพนธ์</a></li>
+                    <li class="nav-item text-center dropdown">
+                        <a class="nav-link dropdown-toggle pe-0" href="#" data-bs-toggle="dropdown">สถิติข้อมูล</a>
+                        <ul class="dropdown-menu dropdown-menu-end">
+                            <li><a class="dropdown-item" href="#">สถิติการจัดเก็บเล่ม</a></li>
+                            <li><a class="dropdown-item" href="/FinalProj/chart_director">สถิติการกำกับเล่ม</a></li>
+                        </ul>
+                    </li>
+                <?php endif; ?>
+
 
             </ul>
         </div>
